+++ frontend/src/components/control/Main.jsx
@@ -1,0 +1,6 @@
+import React from 'react'
+import StyledMain from '../style/MainStyles'
+
+export default function Main () {
+  return (<StyledMain>Home World</StyledMain>)
+}
