+++ frontend/src/components/control/Titlebar.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Container, Logo, PageTitle } from '../style/TitlebarStyles'
+
+export default function Titlebar () {
+  return (
+    <>
+      <Container>
+        <Logo src='#' />
+        <PageTitle>The Bug Amelioration Routine</PageTitle>
+      </Container>
+    </>
+  )
+}
