+++ frontend/src/components/style/Header.jsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+
+export default styled.header`
+  height: 65px;
+  width: 100vw;
+  background: #818797;
+  color: black;
+  border-radius: 0px 0px 0px 5px;
+  border-bottom: 3px solid black;
+  display: grid;
+  grid-template-columns: 92% 8%;
+  justify-items: center;
+  align-items: center;
+`
