+++ frontend/src/components/style/MainStyles.jsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+export default styled.main`
+  margin: 0;
+  background:#382835;
+  color: black;
+  flex-grow: 1;
+  display: grid;
+  grid-template-columns: 60% 40%;
+`
