+++ frontend/src/components/style/TitlebarStyles.jsx
@@ -1,0 +1,19 @@
+import styled from 'styled-components'
+
+export const Container = styled.div`
+  height: 100%;
+  width: 100%;
+  color: black;
+  display: grid;
+  grid-template-columns: 12% 88%;
+  justify-items: center;
+  align-items: center;
+`
+
+export const Logo = styled.img`
+  height: 50px;
+  width: auto;
+  align-self: center;
+`
+
+export const PageTitle = styled.h1``
