+++ frontend/src/components/style/userBox/AuthFormStyles.jsx
@@ -1,0 +1,20 @@
+import styled from 'styled-components'
+
+export const ErrorMsg = styled.p`
+  color: red;
+  font-size: 12px;
+  font-weight: bold;
+`
+
+export const InputLabel = styled.label`
+  margin: 15px 0px 5px;
+  font-weight: bold;
+`
+
+export const AuthBttn = styled.button`
+  color: #DCD7CB;
+  background: #5E809A;
+  height: 40px;
+  width: 80px;
+  margin: 10px;
+`
