+++ frontend/src/components/style/userBox/UserBoxStyles.jsx
@@ -1,0 +1,19 @@
+import styled from 'styled-components'
+
+export const UserBttn = styled.button`
+  background-color: #1E1E21;
+  color: #DCD7CB;
+  font-weight: bold;
+  border-radius: 5px;
+  padding: 5px;
+`
+export const Button = styled.button`
+  color: #DCD7CB;
+  background: #1E1E21;
+  border-radius: 5px;
+  padding: 5px;
+`
+
+export const DisplayUser = styled.h3`
+  color: #DCD7CB;
+`
