+++ frontend/src/index.js
@@ -1,13 +1,11 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './App'
-import Footer from './Components/Footer'
 
 ReactDOM.render(
   <React.StrictMode>
     <>
       <App />
-      <Footer />
     </>
   </React.StrictMode>,
   document.getElementById('root')
